requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
boto3==1.28.42
botocore==1.31.42
celery==5.3.4
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.5
django-cacheops==7.0.1
django-debug-toolbar==4.2.0
django-rest-captcha==0.3.0
django-storages==1.14
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
funcy==1.18
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.3.2
msgpack==1.0.5
now==0.0.6
Pillow==9.5.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.0
s3transfer==0.6.2
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
